KVStore_gRPC Round 1 revise based on code reviews
